Factor python dependency out of interpreter
