+++ torch/csrc/jit/init.cpp
@@ -19,6 +19,7 @@
 #include "torch/csrc/jit/graph_executor.h"
 #include "torch/csrc/jit/script/init.h"
 #include "torch/csrc/jit/script/python_tree_views.h"
+#include "torch/csrc/jit/python_interpreter.h"
 
 
 namespace torch  { namespace jit {
@@ -133,6 +134,7 @@ void initJITBindings(PyObject *module) {
   initPythonTracerBindings(module);
   script::initTreeViewBindings(module);
   script::initJitScriptBindings(module);
+  registerPythonInterpreterOps();
 }
 
 }}
