+++ torch/csrc/jit/interpreter.cpp
@@ -1,6 +1,3 @@
-#ifndef NO_PYTHON
-#include "torch/csrc/python_headers.h"
-#endif
 #include "interpreter.h"
 
 #include "torch/csrc/autograd/edge.h"
@@ -17,17 +14,26 @@
 
 #include <typeinfo>
 
-#ifndef NO_PYTHON
-#include "torch/csrc/autograd/python_engine.h"
-#include "torch/csrc/autograd/python_variable.h"
-#include "torch/csrc/jit/pybind.h"
-#include "torch/csrc/utils/auto_gil.h"
-
-namespace py = pybind11;
-#endif
-
 namespace torch { namespace jit {
 
+
+// externally registered handles, currently used so that python ops
+// can be in a separate compilation unit
+static std::mutex handler_mutex;
+static std::vector<OpHandler> handlers;
+void addInterpreterOpHandler(OpHandler handler) {
+  std::lock_guard<std::mutex> guard(handler_mutex);
+  handlers.push_back(handler);
+}
+at::optional<Operation> lookupExternalOp(Node* n) {
+  std::lock_guard<std::mutex> guard(handler_mutex);
+  for(auto & handler : handlers) {
+    if(auto r = handler(n)) {
+      return *r;
+    }
+  }
+  return at::nullopt;
+}
 
 // Before we translate to intepreter instructions, we do
 // some preprocessing of the graph to turn it into a form that is closer
@@ -374,75 +380,6 @@ bool hasHandleOutput(Node * n) {
   return last->isHandle() && last->uses().size() > 0; // don't bother creating a handle if it is never used
 }
 
-#ifndef NO_PYTHON
-Operation createPythonOperation(PythonOp* op) {
-  py::function func = py::reinterpret_borrow<py::function>(py::handle(op->pyobj.get()));
-  JIT_ASSERT(!hasHandleOutput(op));
-  size_t num_inputs = 0;
-  for(auto arg_type : op->cconv) {
-    if(arg_type == 't')
-      num_inputs++;
-  }
-  return [=](Stack & stack) {
-    AutoGIL gil;
-    py::tuple py_inputs(op->cconv.size());
-    size_t i = 0;
-    size_t next_scalar = 0;
-    size_t next_tensor = 0;
-    for (auto arg_type : op->cconv) {
-      if (arg_type == 's') {
-        py_inputs[i] = py::reinterpret_borrow<py::object>(
-            op->scalar_args[next_scalar++].get());
-      } else if (arg_type == 't') {
-        auto var = peek(stack, next_tensor, num_inputs);
-        py_inputs[i] =
-            py::reinterpret_steal<py::object>(THPVariable_Wrap(var));
-        next_tensor++;
-      }
-      i++;
-    }
-    drop(stack, num_inputs);
-    py::object py_outputs(func(*py_inputs));
-
-    auto num_outputs = op->outputs().size();
-    auto addOutput = [&](py::handle entry) {
-      if (!THPVariable_Check(entry.ptr())) {
-        throw std::runtime_error(
-            "Function application returned a non-Variable output");
-      }
-      THPVariable* var = (THPVariable*)entry.ptr();
-      auto cdata = var->cdata;
-      stack.push_back(std::move(cdata));
-    };
-
-    if (!PyTuple_Check(py_outputs.ptr())) {
-      if (num_outputs != 1) {
-        throw std::runtime_error(
-            "Function.apply returned the wrong number of outputs.");
-      }
-      addOutput(py_outputs);
-    } else {
-      auto output_tuple = py::tuple(py_outputs);
-      if (output_tuple.size() != num_outputs) {
-        throw std::runtime_error(
-            "Function application returned the wrong number of outputs.");
-      }
-      for (py::handle entry : py::tuple(py_outputs)) {
-        addOutput(entry);
-      }
-    }
-    return 0;
-  };
-}
-#else
-Operation createPythonOperation(PythonOp* op) {
-  throw std::runtime_error("Trying to create Python operation from a C++ build");
-  return [=](Stack & stack) {
-    return 0;
-  };
-}
-#endif
-
 Operation createCppOperation(CppOp* op) {
   std::shared_ptr<autograd::Function> func = op->fn;
   JIT_ASSERT(!hasHandleOutput(op));
@@ -464,9 +401,7 @@ Operation createCppOperation(CppOp* op) {
 // Returns a function implementing functionality of a given node,
 // or nullptr if it's a no-op for autograd.
 Operation getOperation(jit::Node* node) {
-  IR_IFM(node, PythonOp)
-    return createPythonOperation(value);
-  IR_ELSEIFM(CppOp)
+  IR_IFM(node, CppOp)
     JIT_ASSERT(!dynamic_cast<autograd::Eval*>(value->fn.get()));
     return createCppOperation(value);
   IR_ELSEIF(FusionGroup)
@@ -588,6 +523,13 @@ Operation getOperation(jit::Node* node) {
       } break;
       default: ;
     };
+
+    // ops registered in another compilation unit
+    // eg: the PythonOp handler
+    if(auto op = lookupExternalOp(node)) {
+      return *op;
+    }
+
     return getTensorOp(node).op;
   IR_END()
 }
