+++ torch/csrc/jit/python_interpreter.h
@@ -1,0 +1,4 @@
+#pragma once
+namespace torch { namespace jit {
+void registerPythonInterpreterOps();
+}}
